+++ .storybook/next-intl.ts
@@ -1,6 +1,7 @@
 import en from '../messages/en.json';
 import de from '../messages/de.json';
-type Messages = typeof en;
+
+type Messages = { en: typeof en; de: typeof de };
 
 const messagesByLocale: Messages = { en, de };