+++ components/commercetools-ui/molecules/info-banner/types/index.ts
@@ -1,0 +1,6 @@
+type Variant = 'primary' | 'warning';
+
+export interface InfoBannerProps {
+  variant?: Variant;
+  className?: string;
+}