+++ components/commercetools-ui/organisms/account/account-atoms/account-dropdown/loggedout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Popover } from '@headlessui/react';
+import { PopoverButton } from '@headlessui/react';
 import { useTranslations } from 'use-intl';
 import Button from 'components/commercetools-ui/atoms/button';
 import Link from 'components/commercetools-ui/atoms/link';
@@ -13,23 +13,23 @@ const LoggedOut = () => {
 
   return (
     <div className="w-235 p-14">
-      <Popover.Button as="div" className="w-full" tabIndex={-1}>
+      <PopoverButton as="div" className="w-full" tabIndex={-1}>
         <Button variant="primary" className="w-full py-12 text-16 leading-[16px]" onClick={goToLoginPage}>
           {translate('account.sign-in')}
         </Button>
-      </Popover.Button>
-      <Popover.Button tabIndex={-1}>
+      </PopoverButton>
+      <PopoverButton tabIndex={-1}>
         <Link link="/" className="mt-20 block w-fit text-primary hover:underline">
           {translate('account.membership-info')}
         </Link>
-      </Popover.Button>
+      </PopoverButton>
       <div className="mt-32 flex pb-10">
         <span className="whitespace-nowrap">{translate('account.not-member')}</span>
-        <Popover.Button tabIndex={-1}>
+        <PopoverButton tabIndex={-1}>
           <Link link="/register" className="ml-4 whitespace-nowrap font-medium text-primary hover:underline">
             {translate('account.join-here')}
           </Link>
-        </Popover.Button>
+        </PopoverButton>
       </div>
     </div>
   );