+++ CHANGELOG.md
@@ -1,19 +1,35 @@
 
-## Version 4.5.0 (2025-06-09)
+## Version 4.6.0 (2025-06-24)
+
+** New Features & Improvements **
+
+- implementation for out of stock items scenarios
+- adding validation message for invalid addresses in checkout
+- fix how colors are formatted in B2C
+- modify code for new color-data
+- Seo information for dynamic pages
+
+
+** Bug Fixes **
+
+- seo meta info bug
+- fixing infinite loop in checkout
+- type issues for storybook
+- Checkout handle out of stock items scenario
+- Filter out inactive shipping methods from display
+
 
 ** New Features & Improvements **
 
 - Upgrade Node to version 24
 - Improved storybook config
 
-
 ** Bug Fixes **
 
-- Fiix price filter
+- Fix price filter
 - Refactor AddressForm to use react-hook-form for form handling
 - Address rendering with improved callback formatting
 - Refactor imports in shipping-section and address-form components
-
 
 ** New Features & Improvements **